check if yum-cron.conf_backup alread exists

+++ centos7minimal-prod.sh
@@ -28,11 +28,18 @@ then
   fi
 fi
 
+# Check if yum-cron.conf_backup already exists
+
+if [ -e /etc/yum/yum-cron.conf_backup ]
+then
 #Configure yum-cron to auto install security updates
 cp -v /etc/yum/yum-cron.conf /etc/yum/yum-cron.conf_backup && \
 sed -i s/update_cmd\ \=\ \default/update_cmd\ \=\ \security/ /etc/yum/yum-cron.conf && \
 sed -i s/apply_updates\ \=\ \no/apply_updates\ \=\ \yes/ /etc/yum/yum-cron.conf
 
+else
+  echo "yum-cron.conf_backup already exists..."
+fi
 #start yum-cron
 systemctl start yum-cron
